Change LAYOUT_SIZE in MBT, Intention, Notification, and Event to return an int instead of a long.

+++ engine/src/main/java/br/angarion/dev/engine/communication/MBT.java
@@ -15,7 +15,7 @@ public final class MBT {
         ValueLayout.JAVA_SHORT.withName("type").withOrder(ByteOrder.BIG_ENDIAN)
     );
 
-    public static final long LAYOUT_SIZE = LAYOUT.byteSize();
+    public static final int LAYOUT_SIZE = (int) LAYOUT.byteSize();
 
     public static final VarHandle TYPE = LAYOUT.varHandle(
         PathElement.groupElement("type")
