+++ engine/src/main/java/br/angarion/dev/engine/communication/notification/Notification.java
@@ -40,7 +40,7 @@ public final class Notification {
         return Short.toUnsignedInt(value);
     }
 
-    public static final long HEADER_SIZE = LAYOUT.byteSize();
+    public static final int HEADER_SIZE = (int) LAYOUT.byteSize();
 
     public static final MemorySegment payloadSlice(MemorySegment message) {
         return message.asSlice(HEADER_SIZE, message.byteSize() - HEADER_SIZE);
